Create 68 freecodecamp
solution to 
Step 68
Call your function passing text and custom_key as the arguments. Store the return value of the function call in a variable called encryption.

solution:- follow it word to word
simply pass function(the one which exists with given arguments)
and return value of "function call" in variable as asked.