Create 81 freecodecamp python
solution to 81 freecodecamp scientifc computing with python

Step 81
The not operator is used to negate an expression. When placed before a truthy value — a value that evaluates to True — it returns False and vice versa.

Add the not operator at the beginning of the if condition to check if the character is not alphabetic.