Create 93 freecodecamp
solution: Step 93
The newline character \n is a special sequence used to represent a new line. You can write a backslash \ followed by an n as a normal sequence of characters in a string and it will be replaced by a new line in the output when the program runs.

Put a newline character at the beginning of your first print call and see the output.