Create freecodecamp 6 lambda
lambda function solution
Step 6
The insert method can add an element at any position in a list. The first argument is the position at which the element has to be added, and the second argument is the element to add. For example, here's how to add a new element in the third position of example_list:

Example Code
example_list = [4, 5, 6, 7]

example_list.insert(2, 5.5)

print(example_list) # [4, 5, 5.5, 6, 7]
Using insert(), add 1 to my_list in the proper position so that it is counting upward, then print the list.