Create 30 lambda function freecodecamp
solution:
Step 30
The filter() function allows you to select items from an iterable, such as a list, based on the output of a function:

Example Code
filter(my_function, my_list)
filter() takes a function as its first argument and an iterable as its second argument. It returns an iterator, which is a special object that enables you to iterate over the elements of a collection, like a list.

The result of the example above is an iterator containing the elements of my_list for which my_function returns True.

Within the filter_expenses_by_category function, call filter() passing the lambda function you wrote in the previous step as the first argument and the expenses list as the second argument.