Create 32 lambda freecode camp
solution:
Step 32
The next step is to define the main function, which will be the entry point of the interactive expense tracker program.

Define a function named main without parameters. Fill the function body with the expenses list you created at the beginning of this project. You will use this list to store the expense records.