Create 43 lambda function freecodecamp
solution:
Step 43
To list all expenses, you can use an elif clause after an if statement. The elif checks additional conditions and only works following an if statement.

Create an elif clause to check if the user's choice equals the string '2'. Inside the elif clause, print the string '\nAll Expenses:'.