Create step 14.py
So far, in this project you have used a for loop to iterate over your input string and convert it into the desired output. Now you'll begin the transition from a for loop to a list comprehension.

Begin by commenting out all the lines of code inside the convert_to_snake_case() function. Don't delete them as they'll be helpful when you implement the logic using a list comprehension.

Remember to add the pass keyword to the function body to prevent the code from failing during the tests.