Create step 18.py
Step 18
After joining the elements of the list snake_cased_char_list, you will need to remove any leading or trailing underscores from the resulting string. For this, use the strip method with the underscore character _ as an argument.

Method calls can be chained together, which means that the result of one method call can be used as the object for another method call.

Example Code
words_list = ['hello', 'world', 'this', 'is', 'chained', 'methods']
result = ' '.join(words_list).upper()
In the example above, the .upper() method is chained to ' '.join(words_list), therefore .upper() is called on the result of the .join() call.