Create step 26.py
Step 26
A tuple is another built-in data structure in Python. Tuples are very much like lists, but they are defined with parentheses (), instead of square brackets. Also, tuples are immutable, unlike lists.

Example Code
my_tuple = ('larch', 1, True)
Your constraints list is going to store tuples. The first item of each tuple will be a constraint parameter.

Modify the constraints list assignment by adding a tuple to your list. Use nums as the first item and an empty string as the second item.