Create step 28.py
Step 28
A regular expression, or regex, is a pattern used to match a specific combination of characters inside a string. It is a valid alternative to if/else conditional statements when you need to match or find patterns inside a string for validation purposes, character replacement, and others.

The compile() function from the re module compiles the string passed as the argument into a regular expression object that can be used by other re methods.

Declare a new pattern variable and assign the value of re.compile('i') to this variable.