Create step 29.py
Step 29
The search() function from the re module analyzes the string passed as the argument looking for the first place where the regex pattern matches the string.

Declare a variable called quote and assign the string 'Not all those who wander are lost.' to this variable. Then, print the result of pattern.search(quote).