Create step 33.py
Step 33
To check that the generated password meets the required features, you are going to use the findall() function from the re module. It's similar to search but it returns a list with all the occurrences of the matched pattern.

Replace the search() call with findall() and check the output.